Refactored Radix Tree Code
