+++ Cargo.toml
@@ -17,4 +17,4 @@ x25519-dalek = "1.2.0"
 opis = "5.1.0"
 sha2 = "0.10.6"
 sha3 = "0.10.5"
-astro-format = "2.0.0"
+astro-format = "2.1.0"
