+++ src/structs/radix_tree.rs
@@ -4,6 +4,12 @@ use astro_format::IntoBytes;
 
 use crate::hash::blake_3;
 
+mod insert;
+mod rehash;
+mod remove;
+mod search;
+mod split_node;
+
 #[derive(Debug,Clone)]
 pub struct RadixTree<K,V> {
     nodes: HashMap<[u8; 32], RadixNode<K,V>>,
@@ -65,7 +71,7 @@ where
 impl<K, V> RadixTree<K, V>
 where
     K: Eq + std::hash::Hash + Clone + std::cmp::Ord + IntoBytes + std::fmt::Display + std::fmt::Debug,
-    V: Clone + IntoBytes,
+    V: Clone + IntoBytes + std::fmt::Debug,
 {
     // Constructor
     pub fn new() -> Self {
@@ -76,357 +82,6 @@ where
         }
     }
 
-    // Insert method
-    pub fn insert<I>(&mut self, key: I, value: V)
-    where I: IntoIterator<Item = K>, {
-
-        let mut current_node_hash = self.root;
-
-        if self.root == [0;32] {
-
-            let key_vec: Vec<K> = key.into_iter().collect();
-
-            let new_node = RadixNode {
-                children: BTreeMap::new(),
-                key: key_vec,
-                value: Some(value.clone()),
-            };
-
-            let new_node_hash = new_node.hash();
-            
-            self.nodes.insert(new_node_hash, new_node);
-
-            self.root = new_node_hash
-
-        } else {
-
-            let mut key_parts = key.into_iter().peekable();
-
-            while let Some(current_key_part) = key_parts.next() {
-
-                let is_last_k = key_parts.peek().is_none();
-
-                let current_node_key_opt = match self.nodes.get(&current_node_hash) {
-                    Some(current_node) => Some(current_node.key.clone()),
-                    None => None,
-                };
-
-                if let Some(current_node_key) = current_node_key_opt {
-
-                    if is_last_k {
-
-                        if let Some(current_node) = self.nodes.get_mut(&current_node_hash) {
-                            current_node.value = Some(value);
-                        }
-
-                        self.rehash(&current_node_hash);
-                        
-                        return;
-
-                    } else {
-
-                        let mut split_position = 0;
-                        let mut matched = true;
-                        
-                        for ck in current_node_key.iter() {
-                            if let Some(next_key_part) = key_parts.next() {
-                                if next_key_part != *ck {
-                                    matched = false;
-                                    break;
-                                }
-                                split_position += 1;
-                            } else {
-                                matched = false;
-                                break;
-                            }
-                        }
-
-                        if matched {
-
-                            if let Some(current_node) = self.nodes.get_mut(&current_node_hash) {
-                                current_node.value = Some(value);
-                            }
-
-                            self.rehash(&current_node_hash);
-                                
-                            return;
-
-                        } else if split_position != 0 {
-
-                            match self.split_node(current_node_hash, split_position) {
-                                Ok((left_node_hash, _right_node_hash)) => {
-                                    current_node_hash = left_node_hash;
-                                },
-                                Err(e) => {
-                                    eprintln!("{}", e);
-                                    return;
-                                }
-                            }
-
-                        } else {
-
-                            let mut new_node_key = vec![];
-                            
-                            if !is_last_k {
-                                while let Some(key_part) = key_parts.next() {
-                                    new_node_key.push(key_part);
-                                }
-                            };
-
-                            let new_node = RadixNode {
-                                children: BTreeMap::new(),
-                                key: new_node_key,
-                                value: Some(value.clone()),
-                            };
-
-                            let new_node_hash = new_node.hash();
-                            
-                            self.nodes.insert(new_node_hash, new_node);
-
-                            if let Some(current_node) = self.nodes.get_mut(&current_node_hash) {
-                                current_node.children.insert(current_key_part, new_node_hash);
-                            }
-
-                            self.parents.insert(new_node_hash, current_node_hash);
-
-                            self.rehash(&current_node_hash);
-
-                        }
-
-                    }
-
-                }
-
-            }
-        
-        }
-
-    }
-
-    
-    // Remove method
-    pub fn remove<I>(&mut self, key: I) -> Result<(), Box<dyn Error>> where I: IntoIterator<Item = K>, {
-        
-        let mut current_node_hash = self.root;
-        
-        let mut key_parts = key.into_iter().peekable();
-
-        while let Some(current_key_part) = key_parts.next() {
-            
-            let is_last_k = key_parts.peek().is_none();
-
-            if let Some(current_node) = self.nodes.get(&current_node_hash) {
-
-                if current_node.key.is_empty() {
-
-                    if is_last_k {
-
-                        if let Some(current_node) = self.nodes.get_mut(&current_node_hash) {
-                            current_node.value = None;
-                        }
-
-                        self.rehash(&current_node_hash);
-
-                        return Ok(());
-
-                    } else {
-
-                        match current_node.children.get(&current_key_part) {
-                            Some(next_node_hash) => {
-                                current_node_hash = *next_node_hash;
-                            },
-                            None => return Ok(()),
-                        }
-
-                    }
-
-                } else {
-
-                    let mut key_matched = true;
-                    let mut key_iter = current_node.key.iter();
-
-                    while let Some(ck) = key_iter.next() {
-                        if Some(ck) != key_parts.next().as_ref() {
-                            key_matched = false;
-                            break;
-                        }
-                    }
-
-                    if key_matched && key_iter.next().is_none() && is_last_k {
-                        if let Some(current_node) = self.nodes.get_mut(&current_node_hash) {
-                            current_node.value = None;
-                        }
-                        self.rehash(&current_node_hash);
-                        return Ok(());
-                    
-                    } else if key_matched {
-
-                        if let Some(next_node_hash) = current_node.children.get(&current_key_part) {
-                            current_node_hash = *next_node_hash;
-                        } else {
-                            return Err("Key does not match!".into());
-                        }
-                    
-                    } else {
-                        return Err("Key does not match!".into());
-                    }
-
-                }
-
-            } else {
-                return Err("Couldn't find node!")?
-            }
-
-        }
-
-        Ok(())
-    }
-
-    pub fn search<I>(&self, key: I) -> Option<&V>
-    where
-        I: IntoIterator<Item = K>,
-    {
-        let mut current_node_hash = &self.root;
-        let mut key_parts = key.into_iter().peekable();
-
-        while let Some(current_key_part) = key_parts.next() {
-            if let Some(node) = self.nodes.get(current_node_hash) {
-                if node.key.is_empty() {
-                    if let Some(next_node_hash) = node.children.get(&current_key_part) {
-                        current_node_hash = next_node_hash;
-                        continue;
-                    } else {
-                        return None;
-                    }
-                } else {
-
-                    let mut key_matched = true;
-
-                    let mut key_iter = node.key.iter();
-
-                    if Some(&current_key_part) != key_iter.next() {
-                        return None;
-                    }
-
-                    while let Some(ck) = key_iter.next() {
-                        if Some(ck) != key_parts.next().as_ref() {
-                            key_matched = false;
-                            break;
-                        }
-                    }
-
-                    if key_matched && key_iter.next().is_none() && key_parts.peek().is_none() {
-                        // The entire key matches and there are no more parts
-                        return node.value.as_ref();
-                    } else if key_matched {
-                        // The key matches so far, continue with the next node
-                        if let Some(next_node_hash) = node.children.get(&key_parts.peek().unwrap_or(&&current_key_part)) {
-                            current_node_hash = next_node_hash;
-                        } else {
-                            return None;
-                        }
-                    } else {
-                        return None;
-                    }
-                }
-            } else {
-                return None;
-            }
-        }
-
-        None
-    }
-
-    fn split_node(&mut self, node_hash: [u8; 32], split_position: usize) -> Result<([u8; 32], [u8; 32]), Box<dyn Error>> {
-        
-        if let Some(old_node) = self.nodes.remove(&node_hash) {
-
-            if split_position < old_node.key.len() {
-                
-                let (left_key_part, right_key_part) = old_node.key.split_at(split_position);
-    
-                let right_node = RadixNode {
-                    key: right_key_part[1..].to_vec(),
-                    children: old_node.children,
-                    value: old_node.value,
-                };
-
-                let right_node_hash = right_node.hash();
-
-                for child_hash in right_node.children.values() {
-                    self.parents.insert(*child_hash, right_node_hash);
-                }
-    
-                self.nodes.insert(right_node_hash, right_node);
-                
-                let mut left_node = RadixNode::new();
-
-                left_node.key = left_key_part.to_vec();
-
-                left_node.children.insert(right_key_part[0].clone(), right_node_hash);
-
-                let left_node_hash = left_node.hash();
-
-                self.nodes.insert(left_node_hash, left_node);
-    
-                // Link the left node to the original node's parent
-                if let Some(parent_hash) = self.parents.remove(&node_hash) {
-                    self.parents.insert(left_node_hash, parent_hash);
-                    if let Some(parent) = self.nodes.get_mut(&parent_hash) {
-                        parent.children.insert(left_key_part[0].clone(), left_node_hash);
-                    }
-                }
-    
-                // Set the parent of the right node to be the new left node
-                self.parents.insert(right_node_hash, left_node_hash);
-
-                return Ok((left_node_hash, right_node_hash))
-            } else {
-                Err("Node can't be split")?
-            }
-        } else {
-            Err("Node doesn't exist!")?
-        }
-    }        
-
-    fn rehash(&mut self, child_hash: &[u8; 32]) {
-
-        let (new_child_hash, parent_hash_opt) = if let Some(child_node) = self.nodes.remove(child_hash) {
-            
-            let new_child_hash = child_node.hash();
-    
-            self.nodes.insert(new_child_hash, child_node);
-
-            let parent_hash_opt = self.parents.get(child_hash).cloned();
-    
-            (new_child_hash, parent_hash_opt)
-
-        } else {
-
-            return;
-
-        };
-
-        if let Some(parent_hash) = parent_hash_opt {
-
-            if let Some(mut parent_node) = self.nodes.remove(&parent_hash) {
-
-                if let Some((child_key, _)) = parent_node.children.iter().find(|(_,v)| *v == child_hash) {
-
-                    parent_node.children.insert(child_key.clone(), new_child_hash);
-
-                }
-    
-                self.nodes.insert(parent_hash, parent_node);
-
-            }
-    
-            self.rehash(&parent_hash);
-
-        }
-
-    }
-    
 }
 
 #[cfg(test)]
@@ -434,16 +89,61 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_insert_and_search() {
-        let mut tree = RadixTree::new();
-        println!("tree.new -> {:?}", tree);
-        let key = vec![0_u8, 1];
-        let value = "value";
-
-        tree.insert(key.clone(), value);
-        println!("tree.insert -> {:?}", tree);
-
-        assert_eq!(tree.search(key), Some(&value));
+    fn test_new() {
+        let tree: RadixTree<String, i32> = RadixTree::new();
+        assert!(tree.nodes.is_empty());
+        assert!(tree.parents.is_empty());
+        // Check if the root is initialized correctly if applicable
     }
 
+    #[test]
+    fn test_insert() {
+        let mut tree = RadixTree::new();
+        tree.insert("key1".as_bytes(), 1);
+        assert_eq!(tree.search("key1".as_bytes()), Some(&1));
+        // Test inserting a key that partially overlaps with an existing key
+        // Test inserting a key that fully overlaps with an existing key
+    }
+
+    #[test]
+    fn test_remove() {
+        let mut tree = RadixTree::new();
+        println!("tree -> {:?}", tree);
+        tree.insert("key1".chars(), 1);
+        println!("tree -> {:?}", tree);
+        assert!(tree.remove("key1".chars()).is_ok());
+        assert_eq!(tree.search("key1".chars()), None);
+        // Test removing a non-existent key
+        // Test removing a key that affects the structure of the tree
+    }
+
+    #[test]
+    fn test_search() {
+        let mut tree = RadixTree::new();
+        tree.insert("key1".as_bytes(), 1);
+        assert_eq!(tree.search("key1".as_bytes()), Some(&1));
+        assert_eq!(tree.search("nonexistent".as_bytes()), None);
+        // Test searching for a key that partially matches an existing key
+    }
+
+    #[test]
+    fn test_node_splitting_with_chars() {
+        let mut tree = RadixTree::new();
+
+        tree.insert("abcde".chars(), 1);
+        
+        tree.insert("abcfg".chars(), 2);
+
+        // Verify that both keys are present and return correct values
+        assert_eq!(tree.search("abcde".chars()), Some(&1));
+        assert_eq!(tree.search("abcfg".chars()), Some(&2));
+
+        // Optionally, verify the absence of a key that should not exist
+        assert_eq!(tree.search("abcf".chars()), None);
+
+        // Additional checks can be made based on the internal structure of your RadixTree
+        assert!(tree.remove("abcde".chars()).is_ok());
+        assert_eq!(tree.search("abcde".chars()), None);
+    }
 }
+
